refactor: extract UI components and standardize utilities
主なリファクタリング内容：

## コンポーネント分離
- TodayDashboardViewから独立したUIコンポーネントを分離（795行→472行）
  - TodoCard, CompletedCard, InProgressBar
  - SwipeableInProgressRow, PriorityBadge
  - SectionContainer, EmptyLabel

## 共通化
- TaskDisplayModelを独立ファイルに移動（Features/Shared/）
- TaskPaletteを作成してカラー定義を統一
- tint関数の重複を削除（ViewModel/View両方で定義されていた）

## タイムゾーン処理の統一
- TimeZone(identifier: "Asia/Tokyo")!を.tokyoに統一
- Date拡張を追加（todayInJST, startOfDayInJST）

## 効果
- コードの重複削減
- ファイルサイズの削減
- 保守性の向上
- 再利用性の向上
