+++ NotionTaskManager/Features/Inbox/InboxCarouselState.swift
@@ -162,7 +162,7 @@ final class InboxCarouselViewModel: ObservableObject {
 
   private static func todayInTokyo() -> Date {
     var calendar = Calendar(identifier: .gregorian)
-    calendar.timeZone = TimeZone(identifier: "Asia/Tokyo")!
+    calendar.timeZone = .tokyo
     return calendar.startOfDay(for: Date())
   }
