+++ NotionTaskManager/Features/Inbox/InboxCarouselView.swift
@@ -387,7 +387,7 @@ struct InboxCardView: View {
 
   private static func todayInTokyo() -> Date {
     var calendar = Calendar(identifier: .gregorian)
-    calendar.timeZone = TimeZone(identifier: "Asia/Tokyo")!
+    calendar.timeZone = .tokyo
     return calendar.startOfDay(for: Date())
   }
 }
