+++ NotionTaskManager/Features/TodayDashboard/Components/EmptyLabel.swift
@@ -1,0 +1,12 @@
+import SwiftUI
+
+struct EmptyLabel: View {
+    let text: String
+
+    var body: some View {
+        Text(text)
+            .font(.subheadline)
+            .foregroundStyle(.secondary)
+            .frame(maxWidth: .infinity, alignment: .leading)
+    }
+}
