+++ NotionTaskManager/Features/TodayDashboard/Components/SectionContainer.swift
@@ -1,0 +1,27 @@
+import SwiftUI
+
+struct SectionContainer<Content: View>: View {
+    let title: String
+    let subtitle: String
+    let tint: Color
+    let foreground: Color
+    @ViewBuilder var content: Content
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 14) {
+            HStack(spacing: 8) {
+                Text(title)
+                    .font(.headline)
+                    .foregroundStyle(foreground)
+                Spacer()
+                Text(subtitle)
+                    .font(.caption)
+                    .foregroundStyle(.secondary)
+            }
+            content
+        }
+        .padding(16)
+        .frame(maxWidth: .infinity, alignment: .leading)
+        .background(tint, in: RoundedRectangle(cornerRadius: 16, style: .continuous))
+    }
+}
